pyproject.toml
[project]
name = "compete-q-qs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.37.33",
    "python-dotenv>=1.1.0",
]
